requirements.txt
altgraph==0.17.4
apimatic-core==0.2.5
apimatic-core-interfaces==0.1.4
apimatic-requests-client-adapter==0.1.6
CacheControl==0.12.14
certifi==2023.7.22
charset-normalizer==3.3.2
deprecation==2.1.0
enum34==1.1.10
idna==3.4
jsonpickle==3.0.2
jsonpointer==2.4
msgpack==1.0.7
mysql-connector==2.2.9
mysql-connector-python==8.2.0
packaging==23.2
patchelf==0.17.2.1
protobuf==4.21.12
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyQt6==6.5.3
PyQt6-Qt6==6.5.3
PyQt6-sip==13.6.0
python-dateutil==2.8.2
pyvin==0.0.2
requests==2.31.0
six==1.16.0
squareup==33.0.0.20231115
urllib3==2.0.7
